finished league score calculator

+++ src/main/java/com/example/basketballgurus/repositories/LeaguesRepository.java
@@ -3,6 +3,6 @@ package com.example.basketballgurus.repositories;
 import com.example.basketballgurus.models.League;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeaguesRepository extends JpaRepository<League, Long > {
+public interface LeaguesRepository extends JpaRepository<League,Integer> {
     
 }
